requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
Django==1.4.3
South==0.7.6


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
mock
django-nose
coverage
django-coverage
selenium
factory_boy==1.2.0
-e git+git://github.com/bitmazk/django-libs.git@8a72ac99ddc27dff22fb06a528db6c81c84b212a#egg=django_libs
django-registration==0.8
